+++ bin/integration-tests/src/tests/client.rs
@@ -7,7 +7,7 @@ use miden_client::{
     keystore::FilesystemKeyStore,
     rpc::domain::account::FetchedAccount,
     store::{InputNoteRecord, InputNoteState, NoteFilter, OutputNoteState, TransactionFilter},
-    testing::common::*,
+    testing::{common::*, config::ClientConfig},
     transaction::{
         DiscardCause, PaymentNoteDescription, TransactionProver, TransactionProverError,
         TransactionRequestBuilder, TransactionStatus,
@@ -21,8 +21,8 @@ use miden_objects::{
 };
 use winter_maybe_async::maybe_async_trait;
 
-pub async fn client_builder_initializes_client_with_endpoint() {
-    let (endpoint, _, store_config, auth_path) = get_client_config();
+pub async fn client_builder_initializes_client_with_endpoint(client_config: ClientConfig) {
+    let (endpoint, _, store_config, auth_path) = client_config.into_parts();
 
     let mut client = ClientBuilder::<FilesystemKeyStore<_>>::new()
         .tonic_rpc_client(&endpoint, Some(10_000))
@@ -40,8 +40,8 @@ pub async fn client_builder_initializes_client_with_endpoint() {
     assert!(sync_summary.block_num.as_u32() > 0);
 }
 
-pub async fn multiple_tx_on_same_block() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn multiple_tx_on_same_block(client_config: ClientConfig) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
     wait_for_node(&mut client).await;
 
     let (first_regular_account, second_regular_account, faucet_account_header) =
@@ -134,12 +134,14 @@ pub async fn multiple_tx_on_same_block() {
     );
 }
 
-pub async fn import_expected_notes() {
-    let (mut client_1, authenticator_1) = create_test_client().await;
+pub async fn import_expected_notes(client_config: ClientConfig) {
+    let (mut client_1, authenticator_1) = create_test_client(client_config.clone()).await;
     let (first_basic_account, faucet_account) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
-    let (mut client_2, authenticator_2) = create_test_client().await;
+    let (mut client_2, authenticator_2) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     let (client_2_account, _seed, _) =
         insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
             .await
@@ -226,15 +228,17 @@ pub async fn import_expected_notes() {
     wait_for_tx(&mut client_1, tx_id).await;
 }
 
-pub async fn import_expected_note_uncommitted() {
-    let (mut client_1, authenticator) = create_test_client().await;
+pub async fn import_expected_note_uncommitted(client_config: ClientConfig) {
+    let (mut client_1, authenticator) = create_test_client(client_config.clone()).await;
     let faucet_account =
         insert_new_fungible_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
             .await
             .unwrap()
             .0;
 
-    let (mut client_2, _) = create_test_client().await;
+    let (mut client_2, _) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     let (client_2_account, _seed, _) =
         insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator)
             .await
@@ -270,12 +274,14 @@ pub async fn import_expected_note_uncommitted() {
     assert!(matches!(imported_note.state(), InputNoteState::Expected { .. }));
 }
 
-pub async fn import_expected_notes_from_the_past_as_committed() {
-    let (mut client_1, authenticator_1) = create_test_client().await;
+pub async fn import_expected_notes_from_the_past_as_committed(client_config: ClientConfig) {
+    let (mut client_1, authenticator_1) = create_test_client(client_config.clone()).await;
     let (first_basic_account, faucet_account) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
-    let (mut client_2, _) = create_test_client().await;
+    let (mut client_2, _) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
 
     wait_for_node(&mut client_2).await;
 
@@ -316,9 +322,9 @@ pub async fn import_expected_notes_from_the_past_as_committed() {
     assert_eq!(imported_note.state(), client_1_note.state());
 }
 
-pub async fn get_account_update() {
+pub async fn get_account_update(client_config: ClientConfig) {
     // Create a client with both public and private accounts.
-    let (mut client, authenticator) = create_test_client().await;
+    let (mut client, authenticator) = create_test_client(client_config.clone()).await;
 
     let (basic_wallet_1, faucet_account) =
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
@@ -350,9 +356,11 @@ pub async fn get_account_update() {
     assert!(matches!(details2, FetchedAccount::Public(_, _)));
 }
 
-pub async fn sync_detail_values() {
-    let (mut client1, authenticator_1) = create_test_client().await;
-    let (mut client2, authenticator_2) = create_test_client().await;
+pub async fn sync_detail_values(client_config: ClientConfig) {
+    let (mut client1, authenticator_1) = create_test_client(client_config.clone()).await;
+    let (mut client2, authenticator_2) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     wait_for_node(&mut client1).await;
     wait_for_node(&mut client2).await;
 
@@ -413,8 +421,10 @@ pub async fn sync_detail_values() {
 
 /// This test runs 3 mint transactions that get included in different blocks so that once we sync
 /// we can check that each transaction gets marked as committed in the corresponding block.
-pub async fn multiple_transactions_can_be_committed_in_different_blocks_without_sync() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn multiple_transactions_can_be_committed_in_different_blocks_without_sync(
+    client_config: ClientConfig,
+) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
 
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
@@ -560,9 +570,11 @@ pub async fn multiple_transactions_can_be_committed_in_different_blocks_without_
 /// - Consuming multiple notes in a single transaction.
 /// - Consuming authenticated notes.
 /// - Consuming unauthenticated notes.
-pub async fn consume_multiple_expected_notes() {
-    let (mut client, authenticator_1) = create_test_client().await;
-    let (mut unauth_client, authenticator_2) = create_test_client().await;
+pub async fn consume_multiple_expected_notes(client_config: ClientConfig) {
+    let (mut client, authenticator_1) = create_test_client(client_config.clone()).await;
+    let (mut unauth_client, authenticator_2) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
 
     wait_for_node(&mut client).await;
 
@@ -646,12 +658,14 @@ pub async fn consume_multiple_expected_notes() {
     }
 }
 
-pub async fn import_consumed_note_with_proof() {
-    let (mut client_1, authenticator_1) = create_test_client().await;
+pub async fn import_consumed_note_with_proof(client_config: ClientConfig) {
+    let (mut client_1, authenticator_1) = create_test_client(client_config.clone()).await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
-    let (mut client_2, authenticator_2) = create_test_client().await;
+    let (mut client_2, authenticator_2) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     let (client_2_account, _seed, _) =
         insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
             .await
@@ -712,13 +726,15 @@ pub async fn import_consumed_note_with_proof() {
     assert!(matches!(consumed_note.state(), InputNoteState::ConsumedExternal { .. }));
 }
 
-pub async fn import_consumed_note_with_id() {
-    let (mut client_1, authenticator) = create_test_client().await;
+pub async fn import_consumed_note_with_id(client_config: ClientConfig) {
+    let (mut client_1, authenticator) = create_test_client(client_config.clone()).await;
     let (first_regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
             .await;
 
-    let (mut client_2, _) = create_test_client().await;
+    let (mut client_2, _) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
 
     wait_for_node(&mut client_2).await;
 
@@ -770,13 +786,15 @@ pub async fn import_consumed_note_with_id() {
     assert!(matches!(consumed_note.state(), InputNoteState::ConsumedExternal { .. }));
 }
 
-pub async fn import_note_with_proof() {
-    let (mut client_1, authenticator) = create_test_client().await;
+pub async fn import_note_with_proof(client_config: ClientConfig) {
+    let (mut client_1, authenticator) = create_test_client(client_config.clone()).await;
     let (first_regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
             .await;
 
-    let (mut client_2, _) = create_test_client().await;
+    let (mut client_2, _) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
 
     wait_for_node(&mut client_2).await;
 
@@ -832,12 +850,14 @@ pub async fn import_note_with_proof() {
     assert!(matches!(imported_note.state(), InputNoteState::Committed { .. }));
 }
 
-pub async fn discarded_transaction() {
-    let (mut client_1, authenticator_1) = create_test_client().await;
+pub async fn discarded_transaction(client_config: ClientConfig) {
+    let (mut client_1, authenticator_1) = create_test_client(client_config.clone()).await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
-    let (mut client_2, authenticator_2) = create_test_client().await;
+    let (mut client_2, authenticator_2) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     let (second_regular_account, ..) =
         insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
             .await
@@ -963,8 +983,8 @@ impl TransactionProver for AlwaysFailingProver {
     }
 }
 
-pub async fn custom_transaction_prover() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn custom_transaction_prover(client_config: ClientConfig) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
 
@@ -996,8 +1016,8 @@ pub async fn custom_transaction_prover() {
     ));
 }
 
-pub async fn locked_account() {
-    let (mut client_1, authenticator) = create_test_client().await;
+pub async fn locked_account(client_config: ClientConfig) {
+    let (mut client_1, authenticator) = create_test_client(client_config.clone()).await;
 
     let (faucet_account, ..) =
         insert_new_fungible_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
@@ -1022,7 +1042,9 @@ pub async fn locked_account() {
     let private_account = client_1.get_account(from_account_id).await.unwrap().unwrap().into();
 
     // Import private account in client 2
-    let (mut client_2, _) = create_test_client().await;
+    let (mut client_2, _) =
+        create_test_client(ClientConfig::default().with_rpc_endpoint(client_config.rpc_endpoint()))
+            .await;
     client_2.add_account(&private_account, seed.into(), false).await.unwrap();
 
     wait_for_node(&mut client_2).await;
@@ -1054,8 +1076,8 @@ pub async fn locked_account() {
     assert!(!account_record.is_locked());
 }
 
-pub async fn expired_transaction_fails() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn expired_transaction_fails(client_config: ClientConfig) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
     let (faucet_account, ..) =
         insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
             .await
@@ -1096,8 +1118,8 @@ pub async fn expired_transaction_fails() {
 
 /// Tests that RPC methods that are not directly related to the client logic
 /// (like GetBlockByNumber) work correctly
-pub async fn unused_rpc_api() {
-    let (mut client, keystore) = create_test_client().await;
+pub async fn unused_rpc_api(client_config: ClientConfig) {
+    let (mut client, keystore) = create_test_client(client_config).await;
 
     let (first_basic_account, faucet_account) =
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Public, &keystore).await;
@@ -1159,8 +1181,8 @@ pub async fn unused_rpc_api() {
     assert_eq!(*account_delta.vault().fungible().iter().next().unwrap().1, MINT_AMOUNT as i64);
 }
 
-pub async fn ignore_invalid_notes() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn ignore_invalid_notes(client_config: ClientConfig) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
     let (regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
             .await;
@@ -1200,8 +1222,8 @@ pub async fn ignore_invalid_notes() {
     assert!(consumed_notes.iter().any(|note| note.id() == note_2.id()));
 }
 
-pub async fn output_only_note() {
-    let (mut client, authenticator) = create_test_client().await;
+pub async fn output_only_note(client_config: ClientConfig) {
+    let (mut client, authenticator) = create_test_client(client_config).await;
 
     let faucet =
         insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)