+++ tests/src/swap_transactions_tests.rs
@@ -1,9 +1,10 @@
 #[tokio::test]
 async fn swap_fully_onchain() {
-    miden_client_integration_tests::tests::swap_transaction::swap_fully_onchain().await
+    miden_client_integration_tests::tests::swap_transaction::swap_fully_onchain(Default::default())
+        .await
 }
 
 #[tokio::test]
 async fn swap_private() {
-    miden_client_integration_tests::tests::swap_transaction::swap_private().await
+    miden_client_integration_tests::tests::swap_transaction::swap_private(Default::default()).await
 }